+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.8.0 - TODO
 
 ### Added
-- Added possibility to specify LTI key and secret through environment variables
+- Added possibility to specify LTI key and secret through environment variables.
 
 ## 1.7.0 - 2020-09-16